Added GUI version of Client
I decided that I wanted to make a more user-friendly version of the client earlier than I was originally planning. This is the GUI; it will eventually have the same functionality as the CLI, but for now, it's still primitive. Right now, it can change the IP manually, set the Host Server (DHCP) IP, and it is password-protected (along with the associated configurations). There's a lightweight installer for all dependencies (right now it's not working, but my next commit will fix it). The client.lua is a helper and is reformatted and slightly changed to support the GUI. The GUI also has access to the CLI version of the client through the helper, so technically, the GUI already has full functionality, but you get it through the CLI window instead of the actual GUI.